libbpf-tools: offcputime: Ensure volatile variable size meets 64-bit requirement
The skeleton code generated by bpftool  uses 'static_assert' to verify
that the size of a volatile variable conforms to that of a 64-bit
system. To satisfy this check, this patch use 'long long' which is 8
bytes on both 32-bit and 64-bit systems.

In eBPF subsystem, the kernel function 'stack_map_alloc' would enforce
8-byte alignment of map size. This patch ensures that
'perf_max_stack_depth' is rounded up to meet the 8-byte alignment
requirement if it is not already aligned.

+++ libbpf-tools/offcputime.bpf.c
@@ -16,7 +16,7 @@ const volatile __u64 max_block_ns = -1;
 const volatile __u64 min_block_ns = 1;
 const volatile bool filter_by_tgid = false;
 const volatile bool filter_by_pid = false;
-const volatile long state = -1;
+const volatile long long state = -1;
 
 struct internal_key {
 	u64 start_ts;
